update exports for vector
