+++ packages/core/src/vector/index.ts
@@ -1,62 +1,2 @@
-import { MastraBase } from '../base';
-import type { Filter } from '../filter';
-
-export interface QueryResult {
-  id: string;
-  score: number;
-  metadata?: Record<string, any>;
-  vector?: number[];
-  /**
-   * The document content, if available.
-   * Note: Currently only supported by Chroma vector store.
-   * For other vector stores, documents should be stored in metadata.
-   */
-  document?: string;
-}
-
-export interface IndexStats {
-  dimension: number;
-  count: number;
-  metric?: 'cosine' | 'euclidean' | 'dotproduct';
-}
-
-export interface UpsertVectorParams {
-  indexName: string;
-  vectors: number[][];
-  metadata?: Record<string, any>[];
-  ids?: string[];
-}
-
-export interface CreateIndexParams {
-  indexName: string;
-  dimension: number;
-  metric?: 'cosine' | 'euclidean' | 'dotproduct';
-}
-
-export type VectorFilter = Filter | null | undefined;
-
-export interface QueryVectorParams {
-  indexName: string;
-  queryVector: number[];
-  topK?: number;
-  filter?: VectorFilter;
-  includeVector?: boolean;
-}
-
-export abstract class MastraVector extends MastraBase {
-  constructor() {
-    super({ name: 'MastraVector', component: 'VECTOR' });
-  }
-
-  abstract upsert(params: UpsertVectorParams): Promise<string[]>;
-
-  abstract createIndex(params: CreateIndexParams): Promise<void>;
-
-  abstract query(params: QueryVectorParams): Promise<QueryResult[]>;
-
-  abstract listIndexes(): Promise<string[]>;
-
-  abstract describeIndex(indexName: string): Promise<IndexStats>;
-
-  abstract deleteIndex(indexName: string): Promise<void>;
-}
+export * from './vector';
+export * from './types';
